refactor(dto): separate request and response sub-packages
