+++ src/main/java/woowacourse/shoppingcart/dto/response/GetMeResponse.java
@@ -1,4 +1,4 @@
-package woowacourse.shoppingcart.dto;
+package woowacourse.shoppingcart.dto.response;
 
 import java.util.Objects;
 import woowacourse.shoppingcart.domain.Customer;
